Rename module dynamic_object to dynamic
