+++ src/kubernetes_api_objects/dynamic.rs
