+++ src/kubernetes_api_objects/mod.rs
@@ -5,7 +5,7 @@ pub mod api_resource;
 pub mod common;
 pub mod config_map;
 pub mod custom_resource;
-pub mod dynamic_object;
+pub mod dynamic;
 pub mod error;
 pub mod object_meta;
 pub mod persistent_volume_claim;
