+++ src/kubernetes_cluster/proof/kubernetes_api_liveness.rs
@@ -1,7 +1,7 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 #![allow(unused_imports)]
-use crate::kubernetes_api_objects::{api_method::*, common::*, dynamic_object::*, error::*};
+use crate::kubernetes_api_objects::{api_method::*, common::*, dynamic::*, error::*};
 use crate::kubernetes_cluster::{
     proof::wf1_assistant::kubernetes_api_action_pre_implies_next_pre,
     spec::{
