+++ src/kubernetes_cluster/spec/kubernetes_api/common.rs
@@ -1,7 +1,7 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 #![allow(unused_imports)]
-use crate::kubernetes_api_objects::{api_method::*, common::*, dynamic_object::*};
+use crate::kubernetes_api_objects::{api_method::*, common::*, dynamic::*};
 use crate::kubernetes_cluster::spec::{channel::*, message::*};
 use crate::state_machine::action::*;
 use crate::state_machine::state_machine::*;
